+++ dev-tools/reconfigurator-cli/Cargo.toml
@@ -21,7 +21,6 @@ humantime.workspace = true
 indexmap.workspace = true
 nexus-inventory.workspace = true
 nexus-reconfigurator-planning.workspace = true
-nexus-reconfigurator-execution.workspace = true
 nexus-sled-agent-shared.workspace = true
 nexus-types.workspace = true
 omicron-common.workspace = true
