+++ nexus/db-queries/src/db/datastore/mod.rs
@@ -444,11 +444,11 @@ mod test {
     use futures::StreamExt;
     use nexus_config::RegionAllocationStrategy;
     use nexus_db_fixed_data::silo::DEFAULT_SILO;
-    use nexus_db_fixed_data::silo::DEFAULT_SILO_ID;
     use nexus_db_model::IpAttachState;
     use nexus_db_model::{to_db_typed_uuid, Generation};
     use nexus_test_utils::db::test_setup_database;
     use nexus_types::external_api::params;
+    use nexus_types::silo::DEFAULT_SILO_ID;
     use omicron_common::api::external::{
         ByteCount, Error, IdentityMetadataCreateParams, LookupType, Name,
     };
@@ -552,8 +552,8 @@ mod test {
         // Associate silo with user
         let authz_silo = authz::Silo::new(
             authz::FLEET,
-            *DEFAULT_SILO_ID,
-            LookupType::ById(*DEFAULT_SILO_ID),
+            DEFAULT_SILO_ID,
+            LookupType::ById(DEFAULT_SILO_ID),
         );
         datastore
             .silo_user_create(
@@ -572,7 +572,7 @@ mod test {
             .fetch()
             .await
             .unwrap();
-        assert_eq!(*DEFAULT_SILO_ID, db_silo_user.silo_id);
+        assert_eq!(DEFAULT_SILO_ID, db_silo_user.silo_id);
 
         // fetch the one we just created
         let (.., fetched) = LookupPath::new(&opctx, &datastore)
@@ -630,7 +630,7 @@ mod test {
             Arc::new(authz::Authz::new(&logctx.log)),
             authn::Context::for_test_user(
                 silo_user_id,
-                *DEFAULT_SILO_ID,
+                DEFAULT_SILO_ID,
                 SiloAuthnPolicy::try_from(&*DEFAULT_SILO).unwrap(),
             ),
             Arc::clone(&datastore) as Arc<dyn nexus_auth::storage::Storage>,
@@ -1726,8 +1726,8 @@ mod test {
         // Create a new Silo user so that we can lookup their keys.
         let authz_silo = authz::Silo::new(
             authz::FLEET,
-            *DEFAULT_SILO_ID,
-            LookupType::ById(*DEFAULT_SILO_ID),
+            DEFAULT_SILO_ID,
+            LookupType::ById(DEFAULT_SILO_ID),
         );
         let silo_user_id = Uuid::new_v4();
         datastore
@@ -1777,7 +1777,7 @@ mod test {
                 .fetch()
                 .await
                 .unwrap();
-        assert_eq!(authz_silo.id(), *DEFAULT_SILO_ID);
+        assert_eq!(authz_silo.id(), DEFAULT_SILO_ID);
         assert_eq!(authz_silo_user.id(), silo_user_id);
         assert_eq!(found.silo_user_id, ssh_key.silo_user_id);
         assert_eq!(found.public_key, ssh_key.public_key);
