+++ nexus/reconfigurator/execution/src/lib.rs
@@ -16,17 +16,13 @@ use nexus_types::deployment::BlueprintZoneFilter;
 use nexus_types::deployment::SledFilter;
 use nexus_types::external_api::views::SledState;
 use nexus_types::identity::Asset;
-use omicron_common::address::Ipv6Subnet;
-use omicron_common::address::SLED_PREFIX;
 use omicron_physical_disks::DeployDisksDone;
 use omicron_uuid_kinds::GenericUuid;
 use omicron_uuid_kinds::OmicronZoneUuid;
 use omicron_uuid_kinds::SledUuid;
-use overridables::Overridables;
 use slog::info;
 use slog_error_chain::InlineErrorChain;
 use std::collections::BTreeMap;
-use std::net::SocketAddrV6;
 use std::sync::Arc;
 use tokio::sync::mpsc;
 use update_engine::merge_anyhow_list;
@@ -37,45 +33,17 @@ mod datasets;
 mod dns;
 mod omicron_physical_disks;
 mod omicron_zones;
-mod overridables;
 mod sagas;
 mod sled_state;
 #[cfg(test)]
 mod test_utils;
 
-pub use dns::blueprint_external_dns_config;
-pub use dns::blueprint_internal_dns_config;
-pub use dns::blueprint_nexus_external_ips;
-pub use dns::silo_dns_name;
-
-pub struct Sled {
-    id: SledUuid,
-    sled_agent_address: SocketAddrV6,
-    is_scrimlet: bool,
-}
-
-impl Sled {
-    pub fn new(
-        id: SledUuid,
-        sled_agent_address: SocketAddrV6,
-        is_scrimlet: bool,
-    ) -> Sled {
-        Sled { id, sled_agent_address, is_scrimlet }
-    }
-
-    pub(crate) fn subnet(&self) -> Ipv6Subnet<SLED_PREFIX> {
-        Ipv6Subnet::<SLED_PREFIX>::new(*self.sled_agent_address.ip())
-    }
-}
-
-impl From<nexus_db_model::Sled> for Sled {
-    fn from(value: nexus_db_model::Sled) -> Self {
-        Sled {
-            id: SledUuid::from_untyped_uuid(value.id()),
-            sled_agent_address: value.address(),
-            is_scrimlet: value.is_scrimlet(),
-        }
-    }
+fn db_sled_to_sled(db_sled: nexus_db_model::Sled) -> Sled {
+    Sled::new(
+        SledUuid::from_untyped_uuid(db_sled.id()),
+        db_sled.address(),
+        db_sled.is_scrimlet(),
+    )
 }
 
 /// The result of calling [`realize_blueprint`] or
@@ -286,7 +254,7 @@ fn register_sled_list_step<'a>(
                     .map(|db_sled| {
                         (
                             SledUuid::from_untyped_uuid(db_sled.id()),
-                            Sled::from(db_sled),
+                            db_sled_to_sled(db_sled),
                         )
                     })
                     .collect();
