+++ nexus/reconfigurator/execution/src/test_utils.rs
@@ -4,13 +4,18 @@
 
 //! Test utilities for reconfigurator execution.
 
+use std::net::Ipv6Addr;
+
 use internal_dns::resolver::Resolver;
 use nexus_db_queries::{context::OpContext, db::DataStore};
-use nexus_types::deployment::{execution::EventBuffer, Blueprint};
+use nexus_types::deployment::{
+    execution::{EventBuffer, Overridables},
+    Blueprint,
+};
 use omicron_uuid_kinds::OmicronZoneUuid;
 use update_engine::TerminalKind;
 
-use crate::{overridables::Overridables, RealizeBlueprintOutput};
+use crate::RealizeBlueprintOutput;
 
 pub(crate) async fn realize_blueprint_and_expect(
     opctx: &OpContext,
@@ -62,4 +67,38 @@ pub(crate) async fn realize_blueprint_and_expect(
     );
 
     (output, buffer)
+}
+
+/// Generates a set of overrides describing the simulated test environment.
+pub fn overridables_for_test(
+    cptestctx: &nexus_test_utils::ControlPlaneTestContext<
+        omicron_nexus::Server,
+    >,
+) -> Overridables {
+    use omicron_common::api::external::SwitchLocation;
+
+    let mut overrides = Overridables::default();
+    let scrimlets = [
+        (nexus_test_utils::SLED_AGENT_UUID, SwitchLocation::Switch0),
+        (nexus_test_utils::SLED_AGENT2_UUID, SwitchLocation::Switch1),
+    ];
+    for (id_str, switch_location) in scrimlets {
+        let sled_id = id_str.parse().unwrap();
+        let ip = Ipv6Addr::LOCALHOST;
+        let mgs_port = cptestctx
+            .gateway
+            .get(&switch_location)
+            .unwrap()
+            .client
+            .bind_address
+            .port();
+        let dendrite_port =
+            cptestctx.dendrite.get(&switch_location).unwrap().port;
+        let mgd_port = cptestctx.mgd.get(&switch_location).unwrap().port;
+        overrides.override_switch_zone_ip(sled_id, ip);
+        overrides.override_dendrite_port(sled_id, dendrite_port);
+        overrides.override_mgs_port(sled_id, mgs_port);
+        overrides.override_mgd_port(sled_id, mgd_port);
+    }
+    overrides
 }
