+++ nexus/src/app/silo.rs
@@ -16,9 +16,9 @@ use nexus_db_queries::db::identity::{Asset, Resource};
 use nexus_db_queries::db::lookup::LookupPath;
 use nexus_db_queries::db::{self, lookup};
 use nexus_db_queries::{authn, authz};
-use nexus_reconfigurator_execution::blueprint_nexus_external_ips;
-use nexus_reconfigurator_execution::silo_dns_name;
+use nexus_types::deployment::execution::blueprint_nexus_external_ips;
 use nexus_types::internal_api::params::DnsRecord;
+use nexus_types::silo::silo_dns_name;
 use omicron_common::api::external::http_pagination::PaginatedBy;
 use omicron_common::api::external::ListResultVec;
 use omicron_common::api::external::LookupResult;
