+++ nexus/tests/integration_tests/disks.rs
@@ -17,7 +17,7 @@ use nexus_db_queries::context::OpContext;
 use nexus_db_queries::db::datastore::RegionAllocationFor;
 use nexus_db_queries::db::datastore::RegionAllocationParameters;
 use nexus_db_queries::db::datastore::REGION_REDUNDANCY_THRESHOLD;
-use nexus_db_queries::db::fixed_data::{silo::DEFAULT_SILO_ID, FLEET_ID};
+use nexus_db_queries::db::fixed_data::FLEET_ID;
 use nexus_db_queries::db::lookup::LookupPath;
 use nexus_test_utils::http_testing::AuthnMode;
 use nexus_test_utils::http_testing::Collection;
@@ -33,6 +33,7 @@ use nexus_test_utils::resource_helpers::objects_list_page_authz;
 use nexus_test_utils::SLED_AGENT_UUID;
 use nexus_test_utils_macros::nexus_test;
 use nexus_types::external_api::params;
+use nexus_types::silo::DEFAULT_SILO_ID;
 use omicron_common::api::external::ByteCount;
 use omicron_common::api::external::Disk;
 use omicron_common::api::external::DiskState;
@@ -1108,7 +1109,7 @@ async fn test_disk_virtual_provisioning_collection(
         0
     );
     let virtual_provisioning_collection = datastore
-        .virtual_provisioning_collection_get(&opctx, *DEFAULT_SILO_ID)
+        .virtual_provisioning_collection_get(&opctx, DEFAULT_SILO_ID)
         .await
         .unwrap();
     assert_eq!(
@@ -1172,7 +1173,7 @@ async fn test_disk_virtual_provisioning_collection(
         0
     );
     let virtual_provisioning_collection = datastore
-        .virtual_provisioning_collection_get(&opctx, *DEFAULT_SILO_ID)
+        .virtual_provisioning_collection_get(&opctx, DEFAULT_SILO_ID)
         .await
         .unwrap();
     assert_eq!(
@@ -1229,7 +1230,7 @@ async fn test_disk_virtual_provisioning_collection(
         disk_size
     );
     let virtual_provisioning_collection = datastore
-        .virtual_provisioning_collection_get(&opctx, *DEFAULT_SILO_ID)
+        .virtual_provisioning_collection_get(&opctx, DEFAULT_SILO_ID)
         .await
         .unwrap();
     assert_eq!(
@@ -1267,7 +1268,7 @@ async fn test_disk_virtual_provisioning_collection(
         0
     );
     let virtual_provisioning_collection = datastore
-        .virtual_provisioning_collection_get(&opctx, *DEFAULT_SILO_ID)
+        .virtual_provisioning_collection_get(&opctx, DEFAULT_SILO_ID)
         .await
         .unwrap();
     assert_eq!(
