+++ nexus/types/src/deployment/execution/overridables.rs
@@ -16,9 +16,9 @@ use std::net::Ipv6Addr;
 ///
 /// Blueprint execution assumes certain values about production systems that
 /// differ in the simulated testing environment and cannot be easily derived
-/// from anything else in the environment.  To accommodate this, this structure
-/// provides access to these values.  Everywhere except the test suite, this
-/// structure is empty and returns the default (production) values.  The test
+/// from anything else in the environment. To accommodate this, this structure
+/// provides access to these values. Everywhere except the test suite, this
+/// structure is empty and returns the default (production) values. The test
 /// suite overrides these values.
 #[derive(Debug, Default)]
 pub struct Overridables {
@@ -34,8 +34,7 @@ pub struct Overridables {
 
 impl Overridables {
     /// Specify the TCP port on which this sled's Dendrite is listening
-    #[cfg(test)]
-    fn override_dendrite_port(&mut self, sled_id: SledUuid, port: u16) {
+    pub fn override_dendrite_port(&mut self, sled_id: SledUuid, port: u16) {
         self.dendrite_ports.insert(sled_id, port);
     }
 
@@ -45,8 +44,7 @@ impl Overridables {
     }
 
     /// Specify the TCP port on which this sled's MGS is listening
-    #[cfg(test)]
-    fn override_mgs_port(&mut self, sled_id: SledUuid, port: u16) {
+    pub fn override_mgs_port(&mut self, sled_id: SledUuid, port: u16) {
         self.mgs_ports.insert(sled_id, port);
     }
 
@@ -56,8 +54,7 @@ impl Overridables {
     }
 
     /// Specify the TCP port on which this sled's MGD is listening
-    #[cfg(test)]
-    fn override_mgd_port(&mut self, sled_id: SledUuid, port: u16) {
+    pub fn override_mgd_port(&mut self, sled_id: SledUuid, port: u16) {
         self.mgd_ports.insert(sled_id, port);
     }
 
@@ -67,8 +64,11 @@ impl Overridables {
     }
 
     /// Specify the IP address of this switch zone
-    #[cfg(test)]
-    fn override_switch_zone_ip(&mut self, sled_id: SledUuid, addr: Ipv6Addr) {
+    pub fn override_switch_zone_ip(
+        &mut self,
+        sled_id: SledUuid,
+        addr: Ipv6Addr,
+    ) {
         self.switch_zone_ips.insert(sled_id, addr);
     }
 
@@ -82,40 +82,5 @@ impl Overridables {
             .get(&sled_id)
             .copied()
             .unwrap_or_else(|| get_switch_zone_address(sled_subnet))
-    }
-
-    /// Generates a set of overrides describing the simulated test environment.
-    #[cfg(test)]
-    pub fn for_test(
-        cptestctx: &nexus_test_utils::ControlPlaneTestContext<
-            omicron_nexus::Server,
-        >,
-    ) -> Overridables {
-        use omicron_common::api::external::SwitchLocation;
-
-        let mut overrides = Overridables::default();
-        let scrimlets = [
-            (nexus_test_utils::SLED_AGENT_UUID, SwitchLocation::Switch0),
-            (nexus_test_utils::SLED_AGENT2_UUID, SwitchLocation::Switch1),
-        ];
-        for (id_str, switch_location) in scrimlets {
-            let sled_id = id_str.parse().unwrap();
-            let ip = Ipv6Addr::LOCALHOST;
-            let mgs_port = cptestctx
-                .gateway
-                .get(&switch_location)
-                .unwrap()
-                .client
-                .bind_address
-                .port();
-            let dendrite_port =
-                cptestctx.dendrite.get(&switch_location).unwrap().port;
-            let mgd_port = cptestctx.mgd.get(&switch_location).unwrap().port;
-            overrides.override_switch_zone_ip(sled_id, ip);
-            overrides.override_dendrite_port(sled_id, dendrite_port);
-            overrides.override_mgs_port(sled_id, mgs_port);
-            overrides.override_mgd_port(sled_id, mgd_port);
-        }
-        overrides
     }
 }
