+++ nexus/types/src/lib.rs
@@ -35,3 +35,4 @@ pub mod external_api;
 pub mod identity;
 pub mod internal_api;
 pub mod inventory;
+pub mod silo;
